+++ src/components/map/MapSigmets.vue
@@ -180,7 +180,7 @@ const geojson = new GeoJSON({
 function buildStyle(color: ColorsList, type: string) {
     return new Style({
         fill: new Fill({
-            color: `rgba(${ getCurrentThemeRgbColor(color).join(',') }, 0.15)`,
+            color: `rgba(${ getCurrentThemeRgbColor(color).join(',') }, ${ store.localSettings.filters?.layers?.transparencySettings?.sigmets || '0.15' })`,
         }),
         stroke: new Stroke({
             color: `rgba(${ getCurrentThemeRgbColor(color).join(',') }, 0.4)`,
@@ -198,7 +198,7 @@ function buildStyle(color: ColorsList, type: string) {
     });
 }
 
-const styles = {
+let styles = {
     default: buildStyle('lightgray125', 'SIGMET'),
     WIND: buildStyle('error300', 'WIND'),
     ICE: buildStyle('primary300', 'ICE'),
@@ -209,10 +209,23 @@ const styles = {
     VA: buildStyle('lightgray125', 'VA'),
 };
 
+watch(() => store.localSettings.filters?.layers?.transparencySettings?.sigmets, () => {
+    styles = {
+        default: buildStyle('lightgray125', 'SIGMET'),
+        WIND: buildStyle('error300', 'WIND'),
+        ICE: buildStyle('primary300', 'ICE'),
+        TURB: buildStyle('warning300', 'TURB'),
+        MTW: buildStyle('warning300', 'MTW'),
+        IFR: buildStyle('info500', 'IFR'),
+        TS: buildStyle('error300', 'TS'),
+        VA: buildStyle('lightgray125', 'VA'),
+    };
+});
+
 function handleMapClick(event: MapBrowserEvent<any>) {
     openSigmet.value = null;
     const features = map.value?.getFeaturesAtPixel(event.pixel, { hitTolerance: 2 });
-    if (!features?.every(x => x instanceof RenderFeature || x.getProperties().dataType || x.getProperties().type === 'local')) return;
+    if (!features?.every(x => x instanceof RenderFeature || x.getProperties().dataType || x.getProperties().type === 'local' || x.getProperties().type === 'root')) return;
 
     const sigmets = features.filter(x => x.getProperties()?.dataType);
     if (!sigmets.length) return;