+++ src/components/views/ViewMapFooter.vue
@@ -70,6 +70,7 @@
                                 Manage friends
                             </common-button>
                             <common-toggle
+                                v-if="store.bookmarks.length"
                                 :model-value="!!store.localSettings.featuredDefaultBookmarks"
                                 @update:modelValue="setUserLocalSettings({ featuredDefaultBookmarks: $event })"
                             >
@@ -362,6 +363,7 @@ function cancelBookingOverride() {
             border-radius: 8px;
 
             font-weight: 300;
+            font-variant-numeric: tabular-nums;
 
             background: $darkgray950;