+++ src/pages/stats/atc.vue
@@ -25,7 +25,7 @@
                 { key: 'actions', name: 'Actions', width: 120 },
             ]"
             item-key="callsign"
-            mobile-width="800px"
+            mobile-width="1000px"
             @click="$event.text_atis && mapStore.addAtcOverlay($event.callsign)"
         >
             <template #data-logon_time="{ item }">