+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+# [1.1.2-final]
+
+## Features and Improvements
+
+- Improved performance when VATGlasses is enabled
+- Added VATSpy/SimAware TRACON geojson debug tool
+- New icons from DotWallop: BCS3, BE35, BE36, E290, F27, F70, F100, HAWK, HDJT, HUSK, L39, M20P, PA18, PA38, SAVG, SU27, TWEN, VIPJ
+- Improved icon for BCS1
+- Added model matching: FH27 -> F27, M20T -> M20, CC19 -> PA18, SU30-32-33-34-35-37 -> SU27, P208 -> TWEN
+- Removed feature positions if they are booked and not actually online
+- Added SIGMETs transparency settings
+- Filter preset save/reset now actually does it's job
+- Applying filter from presets now fully rewrites it's settings instead of merging them
+- You can now configure displayed count or disable airport counters and aircraft labels. Thanks to magicmq for this contribution!
+- Added multiple organisers to event page
+- VATGlasses step is now "500" instead of "1000"
+- Added VATGlasses online-only color support
+- Added diverted flight filter
+
+## Bug Fixes
+
+- Fixed VATGlasses crashing for anonymous users
+- Fixed "A" button disappearing for VG level
+- Fixed last update counter constantly changing its size and causing vG Level to move
+- Fixed an issue when you could not click on SIGMET if non-VG UIR was active
+- Fixed feature position display when no ATIS is online
+- Fixed an issue when VG Combined Mode could get stuck (by Felix)
+- Fixed an error when flight climbing or desending on heading didn't have it's track color change according to new altitude
+- Fixed incorrect VATGlasses automatic altitude assigment for some cases
+- Fixed an issue when you couldn't activate any preset because it always tried to get sorted (touch screens). Sort can now only be done via sorting icon, as it should have been from the start
+
 # [1.1.1]
 
 ## Highlights